Add ERPNext resync with erpnext_id tracking, comment imports, and comprehensive platform updates
- ERPNext importer: switch to get_doc per record for child table access, import HD Ticket
  comments + Communications, Project comments, Task comments with ExternalReference dedup
- Store erpnext_id on Ticket, Project, ProjectTask, Person, Organization for cross-system
  traceability (ERPNext → DotMac Omni → DotMac ERP pipeline)
- Generate DotMac numbers (TKT-/PROJ-/TK- prefixes) during ERPNext import instead of using
  ERPNext document names
- Include erpnext_id in DotMac ERP sync payloads for downstream dedup
- Capture Splynx ID from ERPNext Customer custom fields into subscriber sync_metadata
- Add migration for erpnext_id columns and new ExternalEntityType enum values
- Add 42 tests covering mappers, importer, comment idempotency, and ERP sync payloads
- Drop reseller portal (models, routes, templates, tests)
- Add webhook dead letters, Splynx subscriber pipeline, CRM resolve gate,
  WhatsApp templates, contact merge, live dashboard stats, vendor auth flows
- Platform-wide code quality improvements across all modules

+++ app/services/dem.py
@@ -30,8 +30,8 @@ def _tile_name(latitude: float, longitude: float) -> tuple[str, int, int]:
 
 
 def _tile_indexes(latitude: float, longitude: float, lat_floor: int, lon_floor: int) -> tuple[int, int]:
-    row = int(round((lat_floor + 1 - latitude) * SRTM_SAMPLES_PER_DEGREE))
-    col = int(round((longitude - lon_floor) * SRTM_SAMPLES_PER_DEGREE))
+    row = round((lat_floor + 1 - latitude) * SRTM_SAMPLES_PER_DEGREE)
+    col = round((longitude - lon_floor) * SRTM_SAMPLES_PER_DEGREE)
     row = _clamp(row, 0, SRTM_SAMPLES_PER_DEGREE)
     col = _clamp(col, 0, SRTM_SAMPLES_PER_DEGREE)
     return row, col
